refactor: Change src structure a bit, splitting init & main
